add KnotCompatibilityClassLoader, clean up code, handle CodeSigners/ProtectionDomains better
