+++ src/main/java/net/fabricmc/loader/entrypoint/patches/ModClassLoader_125_FML.java
@@ -21,9 +21,12 @@ import net.fabricmc.loader.util.UrlConversionException;
 import net.fabricmc.loader.util.UrlUtil;
 
 import java.io.File;
+import java.io.InputStream;
 import java.net.MalformedURLException;
 import java.net.URL;
 import java.net.URLClassLoader;
+import java.util.HashSet;
+import java.util.Set;
 
 /**
  * Wrapper class replacing pre-1.3 FML's ModClassLoader (which relies on
@@ -67,5 +70,9 @@ public class ModClassLoader_125_FML extends URLClassLoader {
 	 */
 	public File[] getParentSources() {
 		return new File[] { getParentSource() };
+	}
+
+	static {
+		registerAsParallelCapable();
 	}
 }
