+++ src/main/java/net/fabricmc/loader/launch/knot/KnotClassLoaderInterface.java
@@ -1,0 +1,28 @@
+/*
+ * Copyright 2016 FabricMC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package net.fabricmc.loader.launch.knot;
+
+import java.io.IOException;
+import java.io.InputStream;
+import java.net.URL;
+
+interface KnotClassLoaderInterface {
+	KnotClassDelegate getDelegate();
+	boolean isClassLoaded(String name);
+	void addURL(URL url);
+	InputStream getResourceAsStream(String filename, boolean skipOriginalLoader) throws IOException;
+}
